service worker should also be in ES5
