+++ webpack/service-worker.config.js
@@ -2,6 +2,7 @@ const config = require('sapper/config/webpack.js')
 const terser = require('./terser.config')
 const webpack = require('webpack')
 const { mode, dev, resolve } = require('./shared.config')
+const legacyBabel = require('./legacyBabel.config')
 
 module.exports = {
   entry: config.serviceworker.entry(),
@@ -9,10 +10,16 @@ module.exports = {
   resolve,
   mode,
   devtool: dev ? 'inline-source-map' : 'source-map',
+  module: {
+    rules: [
+      process.env.LEGACY && legacyBabel()
+    ]
+  },
   plugins: [
     new webpack.DefinePlugin({
       'process.browser': true,
       'process.env.NODE_ENV': JSON.stringify(mode),
+      'process.env.LEGACY': !!process.env.LEGACY,
       'process.env.SAPPER_TIMESTAMP': process.env.SAPPER_TIMESTAMP || Date.now()
     }),
     terser()
